- fixed case with shift hotkeys
