+++ packages/ketcher-react/src/script/ui/state/hotkeys.ts
@@ -29,6 +29,7 @@ import {
   runAsyncAction,
   SettingsManager,
   keyNorm,
+  initHotKeys,
 } from 'ketcher-core';
 import { debounce, isEqual } from 'lodash/fp';
 import { load, onAction, removeStructAction } from './shared';
@@ -49,7 +50,7 @@ import { handleHotkeyOverItem } from './handleHotkeysOverItem';
 
 export function initKeydownListener(element) {
   return function (dispatch, getState) {
-    const hotKeys = initHotKeys();
+    const hotKeys = initHotKeys(actions);
     element.addEventListener('keydown', (event) =>
       keyHandle(dispatch, getState, hotKeys, event),
     );
@@ -228,31 +229,6 @@ function getHoveredItem(
   }
 
   return Object.keys(hoveredItem).length ? hoveredItem : null;
-}
-
-function setHotKey(key, actName, hotKeys) {
-  if (Array.isArray(hotKeys[key])) hotKeys[key].push(actName);
-  else hotKeys[key] = [actName];
-}
-
-function initHotKeys() {
-  const hotKeys = {};
-  let act;
-
-  Object.keys(actions).forEach((actName) => {
-    act = actions[actName];
-    if (!act.shortcut) return;
-
-    if (Array.isArray(act.shortcut)) {
-      act.shortcut.forEach((key) => {
-        setHotKey(key, actName, hotKeys);
-      });
-    } else {
-      setHotKey(act.shortcut, actName, hotKeys);
-    }
-  });
-
-  return keyNorm(hotKeys);
 }
 
 function checkGroupOnTool(group, actionTool) {
